+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -533,6 +533,17 @@ All future agent waves must be framed before delegation:
   work, and Task 2.9.4 matrix-row acceptance stay pending. This is local
   package/constants relocation proof only, not runtime proof, AI/hotseat
   product-path support, or procedure-core readiness.
+- Direct-control map/GameInfo constants slice: completed as a narrow public
+  constants ownership relocation. It moves only GameInfo table defaults, map
+  grid bounds, and GameInfo row bounds into `src/play/map/constants.ts` while
+  keeping public facade re-exports in `index.ts`. This preserves map/grid and
+  GameInfo wrapper default and hard-limit behavior plus the capability catalog's
+  injected GameInfo table list. Broader public constants/types, procedure
+  schemas, operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI
+  semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+  matrix-row acceptance stay pending. This is local package/constants relocation
+  proof only, not runtime proof, AI/hotseat product-path support, or
+  procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
