+++ packages/civ7-direct-control/src/play/map/constants.ts
@@ -1,0 +1,19 @@
+export const DEFAULT_CIV7_GAMEINFO_TABLES = [
+  "Resources",
+  "Terrains",
+  "Biomes",
+  "Features",
+  "Units",
+  "UnitOperations",
+  "UnitCommands",
+  "Cities",
+  "CityOperations",
+  "CityCommands",
+  "PlayerOperations",
+  "Maps",
+  "MapSizes",
+] as const;
+export const DEFAULT_CIV7_MAP_GRID_MAX_PLOTS = 512;
+export const HARD_CIV7_MAP_GRID_MAX_PLOTS = 10_000;
+export const DEFAULT_CIV7_GAMEINFO_LIMIT = 100;
+export const HARD_CIV7_GAMEINFO_LIMIT = 1_000;
